Simplify interface; Move tool related interfaces to the tool folder
