+++ web_ui/packages/smart-tools/src/watershed/interfaces.ts
@@ -5,21 +5,13 @@ import { Point } from '../shared/interfaces';
 
 export interface WatershedPolygon {
     id: number;
-    labelId: string;
+    label: { id: string };
     points: Point[];
 }
 
 export type Marker = {
-    labelId: string;
+    label: { id: string };
     points: Point[];
     brushSize: number;
     id: number;
 };
-
-export interface WatershedMethods {
-    executeWatershed: (markers: Marker[], sensitivity: number) => WatershedPolygon[];
-    drawMarkers: (markers: Marker[]) => void;
-    getPolygons: (markers: Marker[]) => WatershedPolygon[];
-    scaleImage: (sensitivity: number) => void;
-    clearMemory: () => void;
-}
